requirements.txt
Django==2.2.6
djangorestframework==3.11.0
celery==4.4.2
psycopg2-binary==2.8.4
